Addition: Attribute validation code

+++ src/character_sheet.rs
@@ -12,13 +12,13 @@ pub struct Template {
 #[derive(Serialize, Deserialize)]
 pub struct Skill {
     pub name: String,
-    pub value: i32
+    pub value: i64
 }
 
 #[derive(Serialize, Deserialize)]
 pub struct Attribute {
     pub name: String,
-    pub value: i32,
+    pub value: i64,
     pub skills: Option<Vec<Skill>>
 }
 
@@ -27,8 +27,8 @@ pub struct CharacterSheet {
     pub name: String,
     pub template: Template,
     pub description: Option<String>,
-    pub health: i32,
-    pub armor_class: i32,
+    pub health: i64,
+    pub armor_class: i64,
     pub weapon_proficiencies: Vec<String>,
     pub perks: Option<Vec<String>>,    
     pub attributes: Vec<Attribute>
